src/main/java/com/matisse/demo/ProductService.java
package com.matisse.demo;

import org.springframework.stereotype.Service;

import java.util.Optional;

@Service
public class ProductService {

    private ProductRepository repository;

    public ProductService(ProductRepository repository) {
        this.repository = repository;
    }

    public Product save(Product product) {
        return repository.save(product);
    }

    public Optional<Product> findById(long aLong) {
        return repository.findById(aLong);
    }

    public boolean existsById(Long aLong) {
        return repository.existsById(aLong);
    }

    public Iterable<Product> findAll() {
        return repository.findAll();
    }

    public Iterable<Product> findAllById(Iterable<Long> iterable) {
        return repository.findAllById(iterable);
    }

    public long count() {
        return repository.count();
    }

    public void deleteById(Long aLong) {
        repository.deleteById(aLong);
    }

    public void delete(Product product) {
        repository.delete(product);
    }

    public void deleteAll(Iterable<Product> iterable) {
        repository.deleteAll(iterable);
    }

    public void deleteAll() {
        repository.deleteAll();
    }
}
